requirements.txt
undetected-chromedriver
selenium
openai
spacy
redis
celery
python-dotenv
requests
transformers